+++ frontend/src/components/MyPage/index.tsx
@@ -1,4 +1,5 @@
 import useProfile from "../../hooks/useProfile";
+import { MediumButton } from "../common/Button";
 import DelProfile from "../kakaoLogin/CharacterDel";
 import CharacterInput from "../kakaoLogin/CharacterInput";
 
@@ -15,10 +16,12 @@ const MyPage = () => {
     <>
       {isLoaded ? (
         isCharacterCheck ? (
-          <DelProfile
-            profileData={profileData}
-            setIsCharacterCheck={setIsCharacterCheck}
-          />
+          <MediumButton
+            variant="contained"
+            onClick={() => DelProfile({ profileData, setIsCharacterCheck })}
+          >
+            계정 삭제
+          </MediumButton>
         ) : (
           <div>
             <CharacterInput
